♻️ [REFAC] #18 차단 번호 조회 API 1차 코드 리뷰 반영
