+++ src/main/java/sopt/appjam/withsuhyeon/controller/AuthController.java
@@ -1,6 +1,5 @@
 package sopt.appjam.withsuhyeon.controller;
 
-import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
